🚚 move funnel things to util
